[player-1788] -moved temp unmute icon styling into its own file -fixed unit tests -added unit tests for unmute icon
