+++ tests/views/adScreen-test.js
@@ -6,6 +6,7 @@ var React = require('react');
 var ReactDOM = require('react-dom');
 var TestUtils = require('react-addons-test-utils');
 var AdScreen = require('../../js/views/adScreen');
+var defaultSkinConfig = require('../../config/skin.json');
 
 describe('AdScreen', function () {
   it('creates an ad screen', function () {
@@ -13,7 +14,10 @@ describe('AdScreen', function () {
     // Render ad screen into DOM
     var mockController = {
       state: {
-        isMobile: false
+        isMobile: false,
+        volumeState: {
+          muted: false
+        }
       }
     };
     var mockSkinConfig = {
@@ -47,7 +51,10 @@ describe('AdScreen', function () {
     var controlBarVisible = true;
     var mockController = {
       state: {
-        isMobile: false
+        isMobile: false,
+        volumeState: {
+          muted: false
+        }
       },
       hideControlBar: function() {
         controlBarVisible = false;
@@ -96,7 +103,10 @@ it('checks that ad marquee is shown/not shown when appropriate', function () {
     var mockController = {
       state: {
         isMobile: false,
-        showAdMarquee: true
+        showAdMarquee: true,
+        volumeState: {
+          muted: false
+        }
       }
     };
     var mockSkinConfig = {
@@ -165,7 +175,10 @@ it('checks that ad marquee is shown/not shown when appropriate', function () {
     var mockController = {
       state: {
         isMobile: false,
-        controlBarVisible: false
+        controlBarVisible: false,
+        volumeState: {
+          muted: false
+        }
       },
       hideControlBar: function() {
         controlBarVisible = false;
@@ -214,7 +227,10 @@ it('checks that ad marquee is shown/not shown when appropriate', function () {
     var mockController = {
       state: {
         isMobile: false,
-        accessibilityControlsEnabled: false
+        accessibilityControlsEnabled: false,
+        volumeState: {
+          muted: false
+        }
       },
       onAdsClicked: function() {
         adsClicked = true;
@@ -256,7 +272,10 @@ it('checks that ad marquee is shown/not shown when appropriate', function () {
       state: {
         isMobile: false,
         accessibilityControlsEnabled: false,
-        controlBarVisible: false
+        controlBarVisible: false,
+        volumeState: {
+          muted: false
+        }
       },
       onAdsClicked: function() {
         adsClicked = true;
@@ -268,7 +287,10 @@ it('checks that ad marquee is shown/not shown when appropriate', function () {
       state: {
         isMobile: true,
         accessibilityControlsEnabled: false,
-        controlBarVisible: true
+        controlBarVisible: true,
+        volumeState: {
+          muted: false
+        }
       },
       onAdsClicked: function() {
         adsClicked = true;
@@ -319,6 +341,50 @@ it('checks that ad marquee is shown/not shown when appropriate', function () {
 
     adScreen.handleTouchEnd(event);
     ReactDOM.unmountComponentAtNode(node);
+  });
+
+  it('should display unmute icon when handling muted autoplay', function () {
+    var mockController = {
+      state: {
+        upNextInfo: {
+          showing: false
+        },
+        volumeState: {
+          muted: true,
+          mutingForAutoplay: true
+        }
+      }
+    };
+
+    var DOM = TestUtils.renderIntoDocument(
+      <AdScreen
+        controller={mockController}
+        skinConfig={defaultSkinConfig}
+      />);
+    var unmuteIcon = TestUtils.findRenderedDOMComponentWithClass(DOM, 'oo-unmute');
+    expect(unmuteIcon).toBeTruthy();
+  });
+
+  it('should not display unmute icon when not muted', function () {
+    var mockController = {
+      state: {
+        upNextInfo: {
+          showing: false
+        },
+        volumeState: {
+          muted: false,
+          mutingForAutoplay: true
+        }
+      }
+    };
+
+    var DOM = TestUtils.renderIntoDocument(
+      <AdScreen
+        controller={mockController}
+        skinConfig={defaultSkinConfig}
+      />);
+    var unmuteIcons = TestUtils.scryRenderedDOMComponentsWithClass(DOM, 'oo-unmute');
+    expect(unmuteIcons.length).toBe(0);
   });
 
 });
